pyproject.toml
[tool.poetry]
name = "utils-api-pipefy"
version = "0.4.1"
description = "Ferramentas para otimizar o consumo paralelizado de api do pipefy, trantando os retornos."
readme = "README.md"
repository = "https://github.com/YuriMotoshima/utils-api-pipefy"
homepage = "https://github.com/YuriMotoshima/utils-api-pipefy"
authors = ["Yuri Motoshima <yurimotoshima@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"GitHub" = "https://github.com/YuriMotoshima/utils-api-pipefy"
